refactor: Use standalone function to get the correct file icon
