+++ src/drive/web/modules/filelist/FileThumbnail.jsx
@@ -5,46 +5,9 @@ import { Icon, withBreakpoints } from 'cozy-ui/react'
 import { ImageLoader } from 'components/Image'
 import { SharedBadge } from 'sharing'
 
-import styles from 'drive/styles/filelist.styl'
-
-import { getFileMimetype } from 'drive/lib/getFileMimetype'
+import getMimeTypeIcon from 'drive/lib/getMimeTypeIcon'
 import { isDirectory } from 'drive/web/modules/drive/files'
-import IconFolder from 'drive/assets/icons/icon-type-folder.svg'
-import IconAudio from 'drive/assets/icons/icon-type-audio.svg'
-import IconBin from 'drive/assets/icons/icon-type-bin.svg'
-import IconCode from 'drive/assets/icons/icon-type-code.svg'
-import IconFiles from 'drive/assets/icons/icon-type-files.svg'
-import IconImage from 'drive/assets/icons/icon-type-image.svg'
-import IconPdf from 'drive/assets/icons/icon-type-pdf.svg'
-import IconSlide from 'drive/assets/icons/icon-type-slide.svg'
-import IconSheet from 'drive/assets/icons/icon-type-sheet.svg'
-import IconText from 'drive/assets/icons/icon-type-text.svg'
-import IconVideo from 'drive/assets/icons/icon-type-video.svg'
-import IconZip from 'drive/assets/icons/icon-type-zip.svg'
-import IconNote from 'drive/assets/icons/icon-type-note.svg'
-
-const getFileIcon = file => {
-  if (isDirectory(file)) {
-    return IconFolder
-  } else if (/\.cozynote$/.test(file.name)) {
-    return IconNote
-  } else {
-    const iconsByMimeType = {
-      audio: IconAudio,
-      bin: IconBin,
-      code: IconCode,
-      image: IconImage,
-      pdf: IconPdf,
-      slide: IconSlide,
-      sheet: IconSheet,
-      text: IconText,
-      video: IconVideo,
-      zip: IconZip
-    }
-    const type = getFileMimetype(iconsByMimeType)(file.mime, file.name)
-    return type ? iconsByMimeType[type] : IconFiles
-  }
-}
+import styles from 'drive/styles/filelist.styl'
 
 const FileThumbnail = ({
   attributes,
@@ -72,7 +35,14 @@ const FileThumbnail = ({
           )}
         />
       ) : (
-        <Icon icon={getFileIcon(attributes)} size={32} />
+        <Icon
+          icon={getMimeTypeIcon(
+            isDirectory(attributes),
+            attributes.name,
+            attributes.mime
+          )}
+          size={32}
+        />
       )}
       {withSharedBadge && (
         <SharedBadge
